:sparkles: Make the cleaning of cluster's identity secrets configurable with default = true (as is)

+++ controllers/vspherecluster_reconciler.go
@@ -202,8 +202,10 @@ func (r clusterReconciler) reconcileDelete(ctx *context.ClusterContext) (reconci
 		if err := ctx.Client.Update(ctx, secret); err != nil {
 			return reconcile.Result{}, err
 		}
-		if err := ctx.Client.Delete(ctx, secret); err != nil {
-			return reconcile.Result{}, err
+		if ctx.CleanClusterIdentitySecrets {
+			if err := ctx.Client.Delete(ctx, secret); err != nil {
+				return reconcile.Result{}, err
+			}
 		}
 	}
